Design Updates, Bug fix, View Addtions
-added Daily view
-moved all quest counting records to record view
-made a components folder
-cleaned up css and separated css for components into respective files
-added descriptions for the questing views so user knows what to do
-updated xp worth for quests
-fixed abandoned quest bug (sort of.. wip, need to redo and update user object in use effect
-brainstorming file added for eggies/leveling up/market place?
