+++ src/components/UserName.css
@@ -1,0 +1,29 @@
+.bodyWrap {
+  background: #e6e3cc;
+}
+
+.welcome {
+  margin-top: 3rem;
+  margin-bottom: 3rem;
+  background: #e6e3cc;
+  text-align: center;
+  font-size: 20px;
+}
+
+.intro {
+  background: #e6e3cc;
+  margin-bottom: 2rem;
+  text-align: center;
+  line-height: 1.5rem;
+  letter-spacing: 0.2px;
+}
+
+.foot {
+  background: #e6e3cc;
+  margin-bottom: 4rem;
+  margin-top: 2rem;
+  text-align: center;
+  line-height: 1.2rem;
+  letter-spacing: 1px;
+  font-size: medium;
+}
