+++ src/text/brainstorm.txt
@@ -1,0 +1,14 @@
+Eggs into leveling ideas
+-Receive an egg on first level up?
+-Receive gold on level up?
+-Purchase further eggs from marketplace
+-purchase items from market place?
+
+
+Market Place-
+(make a marketplace)
+
+Items
+-what kind of items should my dragons need?
+
+Eggs
